pom: Fix unit test failure on GitHub actions (TEDEFO-2937)
The unit tests were failing when run via GitHub Actions because there
was no default country found by Saxon when executing the XSL files.
This caused a DateTimeException when executing the "format-date"
XPath function.

This is probably because the GitHub runners have no locale defined.

Explicitly set the country to BE when running the unit test, via the
Surefire plugin configuration. It has to be done this way because
Surefire forks a new JVM and this needs to be set at the startup of the
JVM. The command line arguments used when running Maven are not passed
along.