Publish via the new Maven Central Portal
Also add the definition of the repository to use for snapshot
dependencies to pom.xml.
Update the definition of this snapshot repository in settings.xml.
The settings.xml file is still needed, as it is used for downloading the
eForms SDK when running the unit tests in the GitHub workflows.

Update README accordingly.